requirements.txt
requests
cvss
boto3
openai
pyyaml